test(st): add cxx-init-test
