+++ test/system_test/cxx_init_test/README.md
@@ -1,0 +1,9 @@
+# cxx_init_test
+
+c++ 构造函数测试
+
+步骤：
+
+1. 在 qemu 中运行
+2. 构建全局变量、静态局部变量
+3. 抛出 throw
